Add the "fill" option.
To load an image, there are aspect & orient option exists.
And here is new option "fill"

it can be "in-size", "over-size"

in-size will fit your image to the size of a part.
over-size will keep the original image size, so if it is bigger than the size of part,
some part of an image will be displayed in your livebox.

Change-Id: I5df8dd329a3be47c69d4126ad0f7dd49bbbc83b6